Fixing selectors members

+++ DataFormats/MuonReco/interface/Muon.h
@@ -226,10 +226,10 @@ namespace reco {
       LowPtMvaMedium         = 1UL<<33,
     };
     
-    bool passed( unsigned int selection ) const { return (selectors_ & selection)==selection; }
+    bool passed( uint64_t selection ) const { return (selectors_ & selection)==selection; }
     bool passed( Selector selection ) const { return passed(static_cast<unsigned int>(selection)); }
-    unsigned int selectors() const { return selectors_; }
-    void setSelectors( unsigned int selectors ){ selectors_ = selectors; }
+    uint64_t selectors() const { return selectors_; }
+    void setSelectors( uint64_t selectors ){ selectors_ = selectors; }
     void setSelector(Selector selector, bool passed){ 
       if (passed)
 	selectors_ |= selector;
